Check existence first when linking to dropbox folder

+++ macos/apps/mountain-duck.sh
@@ -4,4 +4,6 @@
 # Mountain Duck
 ###############################################################################
 
-ln -s /Users/drvdijk/Library/Application\ Support/Mountain\ Duck/Volumes.noindex/Dropbox.localized ~/Dropbox
+if [ ! -e "~/Dropbox" ]; then
+  ln -s /Users/drvdijk/Library/Application\ Support/Mountain\ Duck/Volumes.noindex/Dropbox.localized ~/Dropbox
+fi
